pyproject.toml
[tool.poetry]
name = "cdbot"
version = "0.1.0"
description = "A Discord bot for the Cyber Discovery community."
authors = ["Cyber Discovery Community <triggerhappymods@cyberdiscoverycommunity.uk>"]
license = "MIT"

[tool.poetry.dependencies]
aiohttp = "^3"
"discord.py" = ">=2"
python = "^3.8"
python-dateutil = "^2.8"
asyncpg = "^0.25.0"
sentry-sdk = "^1.5.12"
Pillow = "^8.1"
gitpython = "^3"
httpx = "^0.22.0"
html2markdown = "^0.1.7"
aiocontextvars = "^0.2.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-bandit = "^3"
flake8-bugbear = "^22"
flake8-import-order = "^0.18"
flake8-string-format = "^0.3.0"
flake8-tidy-imports = "^4"
flake8-todo = "^0.7"
flake8-formatter-junit-xml = "^0.0.6"

[tool.poetry.scripts]
cdbot = "cdbot:main"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

